requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
discord.py @ git+https://github.com/Rapptz/discord.py@49e31e9e23517471252cd3f992fa3c88607741ff
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
yarl==1.9.2
